Refactor: extract common code in snapshots/assertions tests generators
