+++ crates/codegen/testing/src/bindings_assertions.rs
@@ -1,5 +1,4 @@
 use std::collections::{BTreeMap, BTreeSet};
-use std::fmt::Write;
 use std::path::Path;
 
 use anyhow::{bail, Result};
@@ -8,24 +7,22 @@ use inflector::Inflector;
 use infra_utils::codegen::CodegenFileSystem;
 use infra_utils::paths::FileWalker;
 
+use crate::common::{generate_mod_file, generate_unit_test_file};
+
 pub fn generate_bindings_assertions_tests(
     language: &Language,
     data_dir: &Path,
     output_dir: &Path,
 ) -> Result<()> {
-    let bindings_tests = collect_bindings_tests(data_dir)?;
+    let tests = collect_assertion_tests(data_dir)?;
 
     let mut fs = CodegenFileSystem::new(data_dir)?;
 
-    generate_mod_file(
-        language,
-        &mut fs,
-        &output_dir.join("mod.rs"),
-        &bindings_tests,
-    )?;
+    generate_mod_file(language, &mut fs, &output_dir.join("mod.rs"), &tests)?;
 
-    for (group_name, test_files) in &bindings_tests {
+    for (group_name, test_files) in &tests {
         generate_unit_test_file(
+            "crate::bindings_assertions::runner",
             &mut fs,
             group_name,
             test_files,
@@ -36,7 +33,7 @@ pub fn generate_bindings_assertions_tests(
     Ok(())
 }
 
-fn collect_bindings_tests(data_dir: &Path) -> Result<BTreeMap<String, BTreeSet<String>>> {
+fn collect_assertion_tests(data_dir: &Path) -> Result<BTreeMap<String, BTreeSet<String>>> {
     let mut bindings_tests = BTreeMap::<String, BTreeSet<String>>::new();
 
     for file in FileWalker::from_directory(data_dir).find(["**/*.sol"])? {
@@ -48,10 +45,11 @@ fn collect_bindings_tests(data_dir: &Path) -> Result<BTreeMap<String, BTreeSet<S
 
         match parts[..] {
             [group_name, test_file] => {
+                let test_name = test_file.strip_suffix(".sol").unwrap();
                 bindings_tests
                     .entry(group_name.to_owned())
                     .or_default()
-                    .insert(test_file.to_owned());
+                    .insert(test_name.to_owned());
             }
             _ => {
                 bail!("Invalid test input. Should be in the form of '<tests-dir>/GROUP_NAME/TEST_FILE.sol', but found: {file:?}");
@@ -60,82 +58,4 @@ fn collect_bindings_tests(data_dir: &Path) -> Result<BTreeMap<String, BTreeSet<S
     }
 
     Ok(bindings_tests)
-}
-
-fn generate_mod_file(
-    language: &Language,
-    fs: &mut CodegenFileSystem,
-    mod_file_path: &Path,
-    bindings_tests: &BTreeMap<String, BTreeSet<String>>,
-) -> Result<()> {
-    let module_declarations_str =
-        bindings_tests
-            .keys()
-            .fold(String::new(), |mut buffer, group_name| {
-                writeln!(buffer, "mod {0};", group_name.to_snake_case()).unwrap();
-                buffer
-            });
-
-    let version_breaks = language.collect_breaking_versions();
-    let version_breaks_len = version_breaks.len();
-    let version_breaks_str = version_breaks
-        .iter()
-        .fold(String::new(), |mut buffer, version| {
-            writeln!(
-                buffer,
-                "Version::new({}, {}, {}),",
-                version.major, version.minor, version.patch
-            )
-            .unwrap();
-            buffer
-        });
-
-    let contents = format!(
-        "
-            use semver::Version;
-            {module_declarations_str}
-
-            pub const VERSION_BREAKS: [Version; {version_breaks_len}] = [
-                {version_breaks_str}
-            ];
-        ",
-    );
-
-    fs.write_file(mod_file_path, contents)
-}
-
-fn generate_unit_test_file(
-    fs: &mut CodegenFileSystem,
-    group_name: &str,
-    test_files: &BTreeSet<String>,
-    unit_test_file_path: &Path,
-) -> Result<()> {
-    let unit_tests_str = test_files
-        .iter()
-        .fold(String::new(), |mut buffer, test_file| {
-            let test_name = test_file.strip_suffix(".sol").unwrap().to_snake_case();
-            writeln!(
-                buffer,
-                r#"
-                    #[test]
-                    fn {test_name}() -> Result<()> {{
-                        run("{group_name}", "{test_file}")
-                    }}
-                "#
-            )
-            .unwrap();
-            buffer
-        });
-
-    let contents = format!(
-        "
-            use anyhow::Result;
-
-            use crate::bindings_assertions::runner::run;
-
-            {unit_tests_str}
-        "
-    );
-
-    fs.write_file(unit_test_file_path, contents)
 }
