+++ crates/codegen/testing/src/lib.rs
@@ -1,5 +1,6 @@
 mod bindings_assertions;
 mod bindings_output;
+mod common;
 mod cst_output;
 
 use std::path::Path;
